+++ core-ng/src/main/java/core/framework/impl/web/http/IPAccessControl.java
@@ -6,22 +6,22 @@ import org.slf4j.LoggerFactory;
 
 import java.net.InetAddress;
 import java.net.UnknownHostException;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+import java.util.stream.Collectors;
 
 /**
  * @author neo
  */
 public class IPAccessControl {
     private final Logger logger = LoggerFactory.getLogger(IPAccessControl.class);
-    private final List<CIDR> cidrs = new ArrayList<>();
+    private final List<CIDR> cidrs;
 
-    public void allowCIDR(String... cidrs) {
-        Arrays.stream(cidrs).map(CIDR::new).forEach(this.cidrs::add);
+    public IPAccessControl(String... cidrs) {
+        this.cidrs = cidrs(cidrs);
     }
 
-    public void validateClientIP(String clientIP) {
+    public void validate(String clientIP) {
         boolean allow = allow(clientIP);
         if (!allow) {
             logger.debug("ip is not from local or match any cidr, cidrs={}", cidrs);
@@ -29,18 +29,9 @@ public class IPAccessControl {
         }
     }
 
-    private boolean allow(String clientIP) {
-        InetAddress address = address(clientIP);
-        if (isLocal(address)) return true;
-        byte[] ipAddress = address.getAddress();
-        for (CIDR cidr : cidrs) {
-            if (cidr.matches(ipAddress)) return true;
-        }
-        return false;
-    }
-
-    boolean isLocal(InetAddress address) {
-        return address.isLoopbackAddress() || address.isSiteLocalAddress();
+    private List<CIDR> cidrs(String... cidrs) {
+        if (cidrs.length == 0) return null;
+        return Arrays.stream(cidrs).map(CIDR::new).collect(Collectors.toList());
     }
 
     private InetAddress address(String address) {
@@ -49,5 +40,21 @@ public class IPAccessControl {
         } catch (UnknownHostException e) {
             throw new Error(e);
         }
+    }
+
+    boolean isLocal(InetAddress address) {
+        return address.isLoopbackAddress() || address.isSiteLocalAddress();
+    }
+
+    private boolean allow(String clientIP) {
+        InetAddress address = address(clientIP);
+        if (isLocal(address)) return true;
+        byte[] ipAddress = address.getAddress();
+        if (cidrs != null) {
+            for (CIDR cidr : cidrs) {
+                if (cidr.matches(ipAddress)) return true;
+            }
+        }
+        return false;
     }
 }