+++ core-ng/src/main/java/core/framework/impl/web/management/APIController.java
@@ -3,7 +3,6 @@ package core.framework.impl.web.management;
 import core.framework.http.ContentType;
 import core.framework.impl.web.api.TypescriptDefinitionBuilder;
 import core.framework.impl.web.http.IPAccessControl;
-import core.framework.util.Lists;
 import core.framework.web.Controller;
 import core.framework.web.Request;
 import core.framework.web.Response;
@@ -14,16 +13,17 @@ import java.util.List;
  * @author neo
  */
 public class APIController implements Controller {
-    public final List<Class<?>> serviceInterfaces = Lists.newArrayList();
+    private final List<Class<?>> serviceInterfaces;
     private final IPAccessControl accessControl;
 
-    public APIController(IPAccessControl accessControl) {
+    public APIController(List<Class<?>> serviceInterfaces, IPAccessControl accessControl) {
+        this.serviceInterfaces = serviceInterfaces;
         this.accessControl = accessControl;
     }
 
     @Override
     public Response execute(Request request) {
-        accessControl.validateClientIP(request.clientIP());
+        accessControl.validate(request.clientIP());
 
         TypescriptDefinitionBuilder builder = new TypescriptDefinitionBuilder();
         serviceInterfaces.forEach(builder::addServiceInterface);