+++ core-ng/src/main/java/core/framework/module/APIConfig.java
@@ -7,7 +7,6 @@ import core.framework.http.HTTPMethod;
 import core.framework.impl.module.ModuleContext;
 import core.framework.impl.web.ControllerHolder;
 import core.framework.impl.web.bean.RequestBeanMapper;
-import core.framework.impl.web.management.APIController;
 import core.framework.impl.web.service.HTTPMethods;
 import core.framework.impl.web.service.WebServiceClient;
 import core.framework.impl.web.service.WebServiceClientBuilder;
@@ -58,7 +57,7 @@ public final class APIConfig {
             }
         }
 
-        apiController().serviceInterfaces.add(serviceInterface);
+        context.serviceInterfaces.add(serviceInterface);
     }
 
     public <T> APIClientConfig client(Class<T> serviceInterface, String serviceURL) {
@@ -104,18 +103,9 @@ public final class APIConfig {
         return state.httpClient;
     }
 
-    private APIController apiController() {
-        if (state.apiController == null) {
-            state.apiController = new APIController(context.httpServer.managementAccessControl);
-            context.route(HTTPMethod.GET, "/_sys/api", state.apiController, true);
-        }
-        return state.apiController;
-    }
-
     public static class State {
         HTTPClient httpClient;
         Duration timeout = Duration.ofSeconds(30);
         Duration slowOperationThreshold = Duration.ofSeconds(15);
-        APIController apiController;
     }
 }