+++ core-ng/src/main/java/core/framework/module/SchedulerConfig.java
@@ -66,7 +66,7 @@ public final class SchedulerConfig {
         context.startupHook.add(scheduler::start);
         context.shutdownHook.add(scheduler::stop);
 
-        SchedulerController schedulerController = new SchedulerController(scheduler, context.httpServer.managementAccessControl);
+        SchedulerController schedulerController = new SchedulerController(scheduler);
         context.route(HTTPMethod.GET, "/_sys/job", schedulerController::jobs, true);
         context.route(HTTPMethod.POST, "/_sys/job/:job", schedulerController::triggerJob, true);