+++ core-ng/src/test/java/core/framework/impl/web/management/ThreadInfoControllerTest.java
@@ -13,7 +13,7 @@ class ThreadInfoControllerTest {
 
     @BeforeEach
     void createThreadInfoController() {
-        controller = new ThreadInfoController(null);
+        controller = new ThreadInfoController();
     }
 
     @Test